Bug Fixes And UI Improvement
bring the title and description above the department?
My profile - On schedule tab it's shows Invalid date.(just a reading error from front end)
"Remove color from subs sidebar.
remove the colors for sub pages"
Employee list - on top of the result set simple show the numbers which is total count of result set that need's to be change as exiting evox.(UI)
Employee List - The icons for view DTR and view Schedules are not coming. (user is also client)
There is 2 My team menu in the nav bar. (one is for supervisor and one is for client, the user could be both super and client)
The department that was first chosen for the announcement should be selected when editing it.
The images were removed while editing the announcement, but they shouldn't have been.